+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+- BLIDE-322 Fix problem interacting with ForgeBox API
 - BLIDE-321 Add `boxlang check`
 
 ## [1.22.3] - 2026-09-14