+++ changelog.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [1.24.1] - 2026-09-19
+
 - BLIDE-325 Fix debugger version tracking
 - BLIDE-322 Fix problem interacting with ForgeBox API
 - BLIDE-321 Add `boxlang check`
@@ -335,7 +337,8 @@
 - Added experimental support for file formatting
 - JSON schema for boxlang.json
 
-[unreleased]: https://github.com/ortus-boxlang/vscode-boxlang/compare/v1.22.3...HEAD
+[unreleased]: https://github.com/ortus-boxlang/vscode-boxlang/compare/v1.24.1...HEAD
+[1.24.1]: https://github.com/ortus-boxlang/vscode-boxlang/compare/v1.22.3...v1.24.1
 [1.22.3]: https://github.com/ortus-boxlang/vscode-boxlang/compare/v1.22.2...v1.22.3
 [1.22.2]: https://github.com/ortus-boxlang/vscode-boxlang/compare/v1.22.1...v1.22.2
 [1.22.1]: https://github.com/ortus-boxlang/vscode-boxlang/compare/v1.22.0...v1.22.1